Moved classes to the right place
