+++ ContactDetailsApi/V1/Domain/Sns/ContactDetailsSns.cs
@@ -15,14 +15,4 @@ namespace ContactDetailsApi.V1.Domain.Sns
         public Guid EntityId { get; set; }
         public EventData EventData { get; set; }
     }
-
-    public class DataItem
-    {
-        public string Value { get; set; }
-    }
-    public class EventData
-    {
-        public DataItem OldData { get; set; }
-        public DataItem NewData { get; set; }
-    }
 }
