+++ ContactDetailsApi/V1/Domain/Sns/DataItem.cs
@@ -1,0 +1,7 @@
+namespace ContactDetailsApi.V1.Domain.Sns
+{
+    public class DataItem
+    {
+        public string Value { get; set; }
+    }
+}
