+++ ContactDetailsApi/V1/Domain/Sns/EventData.cs
@@ -1,0 +1,8 @@
+namespace ContactDetailsApi.V1.Domain.Sns
+{
+    public class EventData
+    {
+        public DataItem OldData { get; set; }
+        public DataItem NewData { get; set; }
+    }
+}
